Refactored to use WebCrawler class
